+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 #### v2.3.2
 - adding ``ResourcePackRegistry``
   - register Built-in resourcepacks from common
+- added more methods to ``ItemModelRenderer``
+  - ``#registerInHandModel``
+  - ``#registerHeadModel``
+  - ``#registerGuiModel``
+  - ``#registerGroundModel``
+  - ``#registerFixedModel``